fix form request

+++ database/migrations/2025_07_04_192429_create_reports_table.php
@@ -21,9 +21,9 @@ return new class extends Migration
             $table->foreignId('completed_by_user_id')->nullable()->constrained('users')->onDelete('set null');
             $table->string('title');
             $table->text('description');
-            $table->string('category')->nullable(); // Pertimbangkan tabel terpisah jika kategori adalah daftar terbatas
-            $table->decimal('latitude', 10, 7)->nullable(); // Sesuaikan presisi sesuai kebutuhan
-            $table->decimal('longitude', 10, 7)->nullable(); // Sesuaikan presisi sesuai kebutuhan
+            $table->string('category')->nullable(); // Pertimbangkan tabel terpisah jika kategori adalah daftar terbatas);
+            $table->float('latitude')->nullable();
+            $table->float('longitude')->nullable();
             $table->text('address')->nullable();
             $table->enum('status', ['pending', 'verified', 'on-progress', 'rejected', 'completed'])->default('pending');
             $table->integer('upvotes_count')->default(0);
